Refactored OrderedSafeExecutor and OrderedScheduler
