+++ bookkeeper-common/src/main/java/org/apache/bookkeeper/common/util/OrderedScheduler.java
@@ -17,32 +17,20 @@
  */
 package org.apache.bookkeeper.common.util;
 
-import static com.google.common.base.Preconditions.checkArgument;
-
 import com.google.common.util.concurrent.ListenableFuture;
 import com.google.common.util.concurrent.ListeningScheduledExecutorService;
-import com.google.common.util.concurrent.ThreadFactoryBuilder;
 
 import io.netty.util.concurrent.DefaultThreadFactory;
 
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.List;
-import java.util.Random;
 import java.util.concurrent.Callable;
-import java.util.concurrent.ExecutionException;
-import java.util.concurrent.Future;
 import java.util.concurrent.ScheduledExecutorService;
 import java.util.concurrent.ScheduledFuture;
 import java.util.concurrent.ScheduledThreadPoolExecutor;
 import java.util.concurrent.ThreadFactory;
+import java.util.concurrent.ThreadPoolExecutor;
 import java.util.concurrent.TimeUnit;
-import java.util.concurrent.TimeoutException;
-import org.apache.bookkeeper.stats.Gauge;
-import org.apache.bookkeeper.stats.NullStatsLogger;
-import org.apache.bookkeeper.stats.OpStatsLogger;
+
 import org.apache.bookkeeper.stats.StatsLogger;
-import org.apache.commons.lang.StringUtils;
 
 /**
  * This class provides 2 things over the java {@link ScheduledExecutorService}.
@@ -58,19 +46,7 @@ import org.apache.commons.lang.StringUtils;
  * achieved by hashing the key objects to threads by their {@link #hashCode()}
  * method.
  */
-public class OrderedScheduler implements ScheduledExecutorService {
-    public static final int NO_TASK_LIMIT = -1;
-    protected static final long WARN_TIME_MICRO_SEC_DEFAULT = TimeUnit.SECONDS.toMicros(1);
-
-    final String name;
-    final ListeningScheduledExecutorService threads[];
-    final long threadIds[];
-    final Random rand = new Random();
-    final OpStatsLogger taskExecutionStats;
-    final OpStatsLogger taskPendingStats;
-    final boolean traceTaskExecution;
-    final long warnTimeMicroSec;
-    final int maxTasksInQueue;
+public class OrderedScheduler extends OrderedSafeExecutor implements ScheduledExecutorService {
 
     /**
      * Create a builder to build ordered scheduler.
@@ -84,61 +60,13 @@ public class OrderedScheduler implements ScheduledExecutorService {
     /**
      * Builder to build ordered scheduler.
      */
-    public static class SchedulerBuilder extends AbstractBuilder<OrderedScheduler> {}
-
-    /**
-     * Abstract builder class to build {@link OrderedScheduler}.
-     */
-    public abstract static class AbstractBuilder<T extends OrderedScheduler> {
-        protected String name = getClass().getSimpleName();
-        protected int numThreads = Runtime.getRuntime().availableProcessors();
-        protected ThreadFactory threadFactory = null;
-        protected StatsLogger statsLogger = NullStatsLogger.INSTANCE;
-        protected boolean traceTaskExecution = false;
-        protected long warnTimeMicroSec = WARN_TIME_MICRO_SEC_DEFAULT;
-        protected int maxTasksInQueue = NO_TASK_LIMIT;
-
-        public AbstractBuilder<T> name(String name) {
-            this.name = name;
-            return this;
-        }
-
-        public AbstractBuilder<T> numThreads(int num) {
-            this.numThreads = num;
-            return this;
-        }
-
-        public AbstractBuilder<T> maxTasksInQueue(int num) {
-            this.maxTasksInQueue = num;
-            return this;
-        }
-
-        public AbstractBuilder<T> threadFactory(ThreadFactory threadFactory) {
-            this.threadFactory = threadFactory;
-            return this;
-        }
-
-        public AbstractBuilder<T> statsLogger(StatsLogger statsLogger) {
-            this.statsLogger = statsLogger;
-            return this;
-        }
-
-        public AbstractBuilder<T> traceTaskExecution(boolean enabled) {
-            this.traceTaskExecution = enabled;
-            return this;
-        }
-
-        public AbstractBuilder<T> traceTaskWarnTimeMicroSec(long warnTimeMicroSec) {
-            this.warnTimeMicroSec = warnTimeMicroSec;
-            return this;
-        }
-
-        @SuppressWarnings("unchecked")
-        public T build() {
+    public static class SchedulerBuilder extends OrderedSafeExecutor.AbstractBuilder<OrderedScheduler> {
+        @Override
+        public OrderedScheduler build() {
             if (null == threadFactory) {
                 threadFactory = new DefaultThreadFactory(name);
             }
-            return (T) new OrderedScheduler(
+            return new OrderedScheduler(
                 name,
                 numThreads,
                 threadFactory,
@@ -146,32 +74,6 @@ public class OrderedScheduler implements ScheduledExecutorService {
                 traceTaskExecution,
                 warnTimeMicroSec,
                 maxTasksInQueue);
-        }
-
-    }
-
-    private class TimedRunnable implements SafeRunnable {
-        final SafeRunnable runnable;
-        final long initNanos;
-
-        TimedRunnable(SafeRunnable runnable) {
-            this.runnable = runnable;
-            this.initNanos = MathUtils.nowInNano();
-         }
-
-        @Override
-        public void safeRun() {
-            taskPendingStats.registerSuccessfulEvent(
-                    MathUtils.elapsedNanos(initNanos),
-                    TimeUnit.NANOSECONDS);
-            long startNanos = MathUtils.nowInNano();
-            this.runnable.safeRun();
-            long elapsedMicroSec = MathUtils.elapsedMicroSec(startNanos);
-            taskExecutionStats.registerSuccessfulEvent(elapsedMicroSec, TimeUnit.MICROSECONDS);
-            if (elapsedMicroSec >= warnTimeMicroSec) {
-                LOGGER.warn("Runnable {}:{} took too long {} micros to execute.",
-                        runnable, runnable.getClass(), elapsedMicroSec);
-            }
         }
     }
 
@@ -191,172 +93,40 @@ public class OrderedScheduler implements ScheduledExecutorService {
      * @param warnTimeMicroSec
      *            - log long task exec warning after this interval
      */
-    protected OrderedScheduler(String baseName,
+    private OrderedScheduler(String baseName,
                                int numThreads,
                                ThreadFactory threadFactory,
                                StatsLogger statsLogger,
                                boolean traceTaskExecution,
                                long warnTimeMicroSec,
                                int maxTasksInQueue) {
-        checkArgument(numThreads > 0);
-        checkArgument(!StringUtils.isBlank(baseName));
-
-        this.maxTasksInQueue = maxTasksInQueue;
-        this.warnTimeMicroSec = warnTimeMicroSec;
-        name = baseName;
-        threads = new ListeningScheduledExecutorService[numThreads];
-        threadIds = new long[numThreads];
-        for (int i = 0; i < numThreads; i++) {
-            final ScheduledThreadPoolExecutor thread = new ScheduledThreadPoolExecutor(1,
-                    new ThreadFactoryBuilder()
-                        .setNameFormat(name + "-" + getClass().getSimpleName() + "-" + i + "-%d")
-                        .setThreadFactory(threadFactory)
-                        .build());
-            threads[i] = new BoundedScheduledExecutorService(thread, this.maxTasksInQueue);
-
-            final int idx = i;
-            try {
-                threads[idx].submit(new SafeRunnable() {
-                    @Override
-                    public void safeRun() {
-                        threadIds[idx] = Thread.currentThread().getId();
-                    }
-                }).get();
-            } catch (InterruptedException e) {
-                Thread.currentThread().interrupt();
-                throw new RuntimeException("Couldn't start thread " + i, e);
-            } catch (ExecutionException e) {
-                throw new RuntimeException("Couldn't start thread " + i, e);
-            }
-
-            // Register gauges
-            statsLogger.registerGauge(String.format("%s-queue-%d", name, idx), new Gauge<Number>() {
-                @Override
-                public Number getDefaultValue() {
-                    return 0;
-                }
-
-                @Override
-                public Number getSample() {
-                    return thread.getQueue().size();
-                }
-            });
-            statsLogger.registerGauge(String.format("%s-completed-tasks-%d", name, idx), new Gauge<Number>() {
-                @Override
-                public Number getDefaultValue() {
-                    return 0;
-                }
-
-                @Override
-                public Number getSample() {
-                    return thread.getCompletedTaskCount();
-                }
-            });
-            statsLogger.registerGauge(String.format("%s-total-tasks-%d", name, idx), new Gauge<Number>() {
-                @Override
-                public Number getDefaultValue() {
-                    return 0;
-                }
-
-                @Override
-                public Number getSample() {
-                    return thread.getTaskCount();
-                }
-            });
-        }
-
-        // Stats
-        this.taskExecutionStats = statsLogger.scope(name).getOpStatsLogger("task_execution");
-        this.taskPendingStats = statsLogger.scope(name).getOpStatsLogger("task_queued");
-        this.traceTaskExecution = traceTaskExecution;
+        super(baseName, numThreads, threadFactory, statsLogger, traceTaskExecution, warnTimeMicroSec, maxTasksInQueue);
     }
 
+
+    @Override
+    protected ScheduledThreadPoolExecutor createSingleThreadExecutor(ThreadFactory factory) {
+        return new ScheduledThreadPoolExecutor(1, factory);
+    }
+
+    @Override
+    protected ListeningScheduledExecutorService getBoundedExecutor(ThreadPoolExecutor executor) {
+        return new BoundedScheduledExecutorService((ScheduledThreadPoolExecutor) executor, this.maxTasksInQueue);
+    }
+
+    @Override
     public ListeningScheduledExecutorService chooseThread() {
-        // skip random # generation in this special case
-        if (threads.length == 1) {
-            return threads[0];
-        }
-
-        return threads[rand.nextInt(threads.length)];
+        return (ListeningScheduledExecutorService) super.chooseThread();
     }
 
+    @Override
     public ListeningScheduledExecutorService chooseThread(Object orderingKey) {
-        // skip hashcode generation in this special case
-        if (threads.length == 1) {
-            return threads[0];
-        }
-
-        return threads[MathUtils.signSafeMod(orderingKey.hashCode(), threads.length)];
+        return (ListeningScheduledExecutorService) super.chooseThread(orderingKey);
     }
 
-    /**
-     * skip hashcode generation in this special case.
-     *
-     * @param orderingKey long ordering key
-     * @return the thread for executing this order key
-     */
+    @Override
     public ListeningScheduledExecutorService chooseThread(long orderingKey) {
-        if (threads.length == 1) {
-            return threads[0];
-        }
-
-        return threads[MathUtils.signSafeMod(orderingKey, threads.length)];
-    }
-
-    private SafeRunnable timedRunnable(SafeRunnable r) {
-        if (traceTaskExecution) {
-            return new TimedRunnable(r);
-        } else {
-            return r;
-        }
-    }
-
-    /**
-     * schedules a one time action to execute.
-     */
-    public void submit(SafeRunnable r) {
-        chooseThread().submit(timedRunnable(r));
-    }
-
-    /**
-     * schedules a one time action to execute with an ordering guarantee on the <tt>orderingKey</tt>.
-     *
-     * @param orderingKey order key to submit the task
-     * @param r task to run
-     * @return listenable future on the completion of the task
-     */
-    public ListenableFuture<?> submitOrdered(Object orderingKey, SafeRunnable r) {
-        return chooseThread(orderingKey).submit(timedRunnable(r));
-    }
-
-    /**
-     * schedules a one time action to execute with an ordering guarantee on the <tt>orderingKey</tt>.
-     *
-     * @param orderingKey order key to submit the task
-     * @param r task to run
-     */
-    public void executeOrdered(Object orderingKey, SafeRunnable r) {
-        chooseThread(orderingKey).execute(timedRunnable(r));
-    }
-
-    /**
-     * schedules a one time action to execute with an ordering guarantee on the key.
-     *
-     * @param orderingKey
-     * @param r
-     */
-    public void submitOrdered(long orderingKey, SafeRunnable r) {
-        chooseThread(orderingKey).execute(timedRunnable(r));
-    }
-
-    /**
-     * schedules a one time action to execute with an ordering guarantee on the key.
-     *
-     * @param orderingKey
-     * @param r
-     */
-    public void submitOrdered(int orderingKey, SafeRunnable r) {
-        chooseThread(orderingKey).execute(timedRunnable(r));
+        return (ListeningScheduledExecutorService) super.chooseThread(orderingKey);
     }
 
     /**
@@ -472,91 +242,6 @@ public class OrderedScheduler implements ScheduledExecutorService {
         return chooseThread(orderingKey).scheduleWithFixedDelay(command, initialDelay, delay, unit);
     }
 
-    protected long getThreadID(long orderingKey) {
-        // skip hashcode generation in this special case
-        if (threadIds.length == 1) {
-            return threadIds[0];
-        }
-
-        return threadIds[MathUtils.signSafeMod(orderingKey, threadIds.length)];
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public void shutdown() {
-        for (int i = 0; i < threads.length; i++) {
-            threads[i].shutdown();
-        }
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public List<Runnable> shutdownNow() {
-        List<Runnable> runnables = new ArrayList<Runnable>();
-        for (ScheduledExecutorService executor : threads) {
-            runnables.addAll(executor.shutdownNow());
-        }
-        return runnables;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public boolean isShutdown() {
-        for (ScheduledExecutorService executor : threads) {
-            if (!executor.isShutdown()) {
-                return false;
-            }
-        }
-        return true;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public boolean awaitTermination(long timeout, TimeUnit unit) throws InterruptedException {
-        boolean ret = true;
-        for (int i = 0; i < threads.length; i++) {
-            ret = ret && threads[i].awaitTermination(timeout, unit);
-        }
-        return ret;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public boolean isTerminated() {
-        for (ScheduledExecutorService executor : threads) {
-            if (!executor.isTerminated()) {
-                return false;
-            }
-        }
-        return true;
-    }
-
-    /**
-     * Force threads shutdown (cancel active requests) after specified delay,
-     * to be used after shutdown() rejects new requests.
-     */
-    public void forceShutdown(long timeout, TimeUnit unit) {
-        for (int i = 0; i < threads.length; i++) {
-            try {
-                if (!threads[i].awaitTermination(timeout, unit)) {
-                    threads[i].shutdownNow();
-                }
-            } catch (InterruptedException exception) {
-                threads[i].shutdownNow();
-                Thread.currentThread().interrupt();
-            }
-        }
-    }
 
     //
     // Methods for implementing {@link ScheduledExecutorService}
@@ -594,76 +279,6 @@ public class OrderedScheduler implements ScheduledExecutorService {
     public ScheduledFuture<?> scheduleWithFixedDelay(Runnable command,
                                                      long initialDelay, long delay, TimeUnit unit) {
         return chooseThread().scheduleWithFixedDelay(command, initialDelay, delay, unit);
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public <T> Future<T> submit(Callable<T> task) {
-        return chooseThread().submit(task);
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public <T> Future<T> submit(Runnable task, T result) {
-        return chooseThread().submit(task, result);
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public Future<?> submit(Runnable task) {
-        return chooseThread().submit(task);
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public <T> List<Future<T>> invokeAll(Collection<? extends Callable<T>> tasks)
-        throws InterruptedException {
-        return chooseThread().invokeAll(tasks);
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public <T> List<Future<T>> invokeAll(Collection<? extends Callable<T>> tasks,
-                                         long timeout,
-                                         TimeUnit unit)
-        throws InterruptedException {
-        return chooseThread().invokeAll(tasks, timeout, unit);
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public <T> T invokeAny(Collection<? extends Callable<T>> tasks)
-        throws InterruptedException, ExecutionException {
-        return chooseThread().invokeAny(tasks);
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public <T> T invokeAny(Collection<? extends Callable<T>> tasks, long timeout, TimeUnit unit)
-        throws InterruptedException, ExecutionException, TimeoutException {
-        return chooseThread().invokeAny(tasks, timeout, unit);
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public void execute(Runnable command) {
-        chooseThread().execute(command);
     }
 
 }
